PythonFlaskServer/requirements.txt
click==8.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
zipp==3.8.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
python-socketio==4.0.0
python-engineio==3.0.0
flask_socketio==4.1.0
flask==2.1.3
werkzeug==2.0.0